+++ tests/testthat/test-label.R
@@ -106,9 +106,9 @@ test_that("label", {
   m_explicit <- mlabel(stimuli, text = c("A", "B"))
   gg_explicit <- gglabel(stimuli, label = c("A", "B"))
   
-  same_m <- image_comp(m_guess[1], m_explicit[1])
-  same_gg <- image_comp(gg_guess[1], gg_explicit[1])
-  diff <- image_comp(gg_guess[1], m_explicit[1])
+  same_m <- compare(m_guess[1], m_explicit[1])
+  same_gg <- compare(gg_guess[1], gg_explicit[1])
+  diff <- compare(gg_guess[1], m_explicit[1])
   
   expect_equal(same_m[[1]], 0)
   expect_equal(same_gg[[1]], 0)